% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_postprocessing.R
\name{calculate_CF_age}
\alias{calculate_CF_age}
\title{calculate CF by age
Takes the output from infectionsCF and computes the age}
\usage{
calculate_CF_age(
  fred_key,
  fred_n,
  brk_ages_cf_in = c(seq(from = 0, by = 10, to = 80), 120),
  save_file = FALSE
)
}
\arguments{
\item{fred_key}{key of fred job}

\item{fred_n}{replicate of the FRED job to calculate R0 over}

\item{save_file}{should it save results to a file?}

\item{age_brks}{breaks for ages}
}
\value{
a list of dataframes: periods has the inf and symp periods, intervals the serial intervals
}
\description{
calculate CF by age
Takes the output from infectionsCF and computes the age
}
\examples{
calculate_CF_age('test_cf_age', 1, seq(from=0,by=10,to=90))
}
