% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_postprocessing.R
\name{calculate_CF_intervals}
\alias{calculate_CF_intervals}
\title{calculate CF intervals
calculate delays of CF from symptoms and infection}
\usage{
calculate_CF_intervals(fred_key, fred_n)
}
\arguments{
\item{fred_key}{key of fred job}

\item{fred_n}{replicate of the FRED job to calculate R0 over}
}
\value{
a list of dataframes: periods has the inf and symp periods, intervals the serial intervals
}
\description{
calculate CF intervals
calculate delays of CF from symptoms and infection
}
\examples{
calculate_CF_intervals('test_cf_intervals', 1)
}
