% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_postprocessing.R
\name{calculate_fred_R0}
\alias{calculate_fred_R0}
\title{calculate_fred_R0}
\usage{
calculate_fred_R0(fred_key, fred_n)
}
\arguments{
\item{fred_key}{key of the FRED job}

\item{fred_n}{number of replicates in FRED job}
}
\value{
doesn't return anything. It writes the output to two files
}
\description{
Calculate the R(t) for a FRED_JOB
}
\examples{
calculate_fred_R0(test_job, 10)
}
