% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_postprocessing.R
\name{calculate_intervals}
\alias{calculate_intervals}
\title{calculate intervals
calculate serial intervals and infectious periods}
\usage{
calculate_intervals(fred_key, fred_n)
}
\arguments{
\item{fred_key}{key of fred job}

\item{fred_n}{replicate of the FRED job to calculate R0 over}
}
\value{
a list of dataframes: periods has the inf and symp periods, intervals the serial intervals
}
\description{
calculate intervals
calculate serial intervals and infectious periods
}
\examples{
calculate_intervals('FRED_RESULTS/JOB/1/DATA/OUT/', 1)
}
