% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_postprocessing.R
\name{calculate_R0}
\alias{calculate_R0}
\title{calculate_R0}
\usage{
calculate_R0(data_dir_in, n_in)
}
\arguments{
\item{data_dir_in}{directory where the output is}

\item{n_in}{replicate of the FRED job to calculate R0 over}
}
\value{
returns a timeseries of R(t)
}
\description{
Calculate the R(t) for a FRED_JOB with data folder specified and for one run
}
\examples{
calculate_R0('FRED_RESULTS/JOB/1/DATA/OUT/', 1)
}
