% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_control_files.R
\name{fred_gather_data}
\alias{fred_gather_data}
\title{fred_gather_data}
\usage{
fred_gather_data(
  params,
  outdir,
  outfile,
  FUN,
  FUN2 = function(x) {     return(x) },
  rm_out = FALSE
)
}
\arguments{
\item{params}{parameters of the simulation}

\item{outdir}{output directory}

\item{FUN}{function to check if simulations finished}

\item{FUN2}{function to post-process data}

\item{rm_out}{shuold remove output directory if exists?}
}
\value{
returns 0 if things worked fine
}
\description{
Collects data from fred_jobs
}
\examples{
fred_gather_data("parameters.csv", "FRED_Sims_out")
}
