% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_control_files.R
\name{submit_jobs}
\alias{submit_jobs}
\title{submits jobs to cluster
This function takes specific parameters and creates the submission script that can be used later on to submit a job array of simulations}
\usage{
submit_jobs(experiment_supername_in, experiment_name_in, experiment_dir_in,
  params_base, job_base, reps, scalars, FUN, cores_in = 1,
  delete_files = F, sub_array = T)
}
\arguments{
\item{experiment_supername_in}{input the main name of the experiment, assuming there will be subexperiments for each submission}

\item{experiment_name_in}{input the name of the main experiment, would probably be the same as experiment_dir_in, except the experiment_dir_in can have a different path}

\item{experiment_dir_in}{input is the name of the main directory where the simulations will be located}

\item{params_base}{prefix of parameters file}

\item{job_base}{prefix of job key}

\item{reps}{repetitions for each job}

\item{delete_files}{(TRUE or FALSE) delete files after submission}

\item{n}{total simulations}
}
\value{
a .sh file that is written to memory
}
\description{
submits jobs to cluster
This function takes specific parameters and creates the submission script that can be used later on to submit a job array of simulations
}
\examples{

submit_jobs(
   executable_path_in = executable_path,
   input_files_path_in = input_files_path,
   experiment_supername_in = experiment_supername,
   experiment_name_in = experiment_name,
   experiment_dir_in = experiment_dir,params_base,job_base,reps,n,delete_files = T)
}
