% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_control_files.R
\name{write_fred_parameters}
\alias{write_fred_parameters}
\title{write_fred_parameters}
\usage{
write_fred_parameters(
  scalars.in,
  defaults.file,
  dir.in,
  basename.in = "fred_params",
  fred_defaults = "./FRED/input_files/defaults"
)
}
\arguments{
\item{scalars.in}{data frame with parameters}

\item{defaults.file}{file with default parameters}

\item{dir.in}{directory where the simulation files will be saved}

\item{basename.in}{basename for the file names}
}
\value{
returns 0 if ERROR
}
\description{
write_fred_parameters
}
\examples{
write_fred_parameters(scalars, 'params.txt', 'simfiles','fred_params')
}
